h07 operations

+++ h07/src/libread/read.c
@@ -18,10 +18,10 @@ int read_string(char *str) {
   }
   return 0;
 }
-int read_number(int *number) {
+int read_number(float *number) {
   char buff[STR_LEN];
   if (fgets(buff, sizeof(buff), stdin)) {
-    if (sscanf(buff, "%d", number) == 0) {
+    if (sscanf(buff, "%f", number) == 0) {
       printf("can't read number!\n");
       return 1;
     }
