+++ h07/src/libread/read.h
@@ -4,4 +4,4 @@
 #define STR_LEN 10
 
 int read_string(char *string);
-int read_number(int *number);
+int read_number(float *number);
